Move PSF3D tests
